AI_Form_Filling_Assistance/requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
pillow>=9.0.0
pdf2image>=1.16.0
PyMuPDF>=1.21.0
easyocr>=1.7.0
